+++ src/core/src/lib/components/formly.attributes.spec.ts
@@ -98,11 +98,11 @@ describe('FormlyAttributes Component', () => {
       const fixture = createTestComponent('<input type="text" [formlyAttributes]="field">');
       const elm = getFormlyAttributesElement(fixture.nativeElement);
 
-      fixture.componentInstance.field = { focus: true, templateOptions: {} };
+      fixture.componentInstance.field.focus = true;
       fixture.detectChanges();
       expect(document.activeElement === elm).toBeTruthy();
 
-      fixture.componentInstance.field = { focus: false, templateOptions: {} };
+      fixture.componentInstance.field.focus = false;
       fixture.detectChanges();
       expect(document.activeElement === elm).toBeFalsy();
     });
@@ -116,6 +116,19 @@ describe('FormlyAttributes Component', () => {
 
       directive.triggerEventHandler('blur', {});
       expect(fixture.componentInstance.field.focus).toBeFalsy();
+    });
+
+    it(`should focus the first element when mutliple formlyAttributes is present`, () => {
+      const fixture = createTestComponent(`
+        <input type="text" [formlyAttributes]="field">
+        <input type="text" [formlyAttributes]="field">
+        <input type="text" [formlyAttributes]="field">
+      `);
+      const elm = getFormlyAttributesElement(fixture.nativeElement);
+
+      fixture.componentInstance.field.focus = true;
+      fixture.detectChanges();
+      expect(document.activeElement === elm).toBeTruthy();
     });
   });
 });