+++ src/core/src/lib/extensions/field-form/field-form.ts
@@ -1,7 +1,7 @@
 import { FormlyExtension } from '../../services/formly.config';
 import { FormlyFieldConfigCache } from '../../components/formly.field.config';
 import { AbstractControl, FormGroup, FormArray, FormControl, AbstractControlOptions } from '@angular/forms';
-import { getKeyPath, isNullOrUndefined } from '../../utils';
+import { getKeyPath, isNullOrUndefined, defineHiddenProp, assignModelValue, getFieldValue } from '../../utils';
 
 /** @experimental */
 export class FieldFormExtension implements FormlyExtension {
@@ -17,7 +17,7 @@ export class FieldFormExtension implements FormlyExtension {
           this.addFormControl(rootForm, field, rootModel, formPath);
         } else {
           if (!rootModel[path]) {
-            rootModel[path] = typeof path === 'string' ? {} : [];
+            assignModelValue(rootModel, field.key, getFieldValue(field));
           }
           this.addFormControl(rootForm, { key: formPath, fieldGroup: [], modelOptions: {}, templateOptions: {} }, rootModel, formPath);
 
@@ -33,6 +33,10 @@ export class FieldFormExtension implements FormlyExtension {
   }
 
   private addFormControl(form: FormGroup | FormArray, field: FormlyFieldConfigCache, model: any, path: string | number) {
+    if (!field.formControl) {
+      defineHiddenProp(field, 'formControl', field.formControl);
+    }
+
     const abstractControlOptions = {
       validators: field._validators,
       asyncValidators: field._asyncValidators,