+++ src/core/src/lib/services/formly.form.builder.ts
@@ -1,30 +1,20 @@
-import { Injectable, ComponentFactoryResolver } from '@angular/core';
+import { Injectable, ComponentFactoryResolver, Injector } from '@angular/core';
 import { FormGroup, FormArray } from '@angular/forms';
 import { FormlyConfig } from './formly.config';
 import { FormlyFieldConfig, FormlyFormOptions, FormlyFieldConfigCache, FormlyValueChangeEvent, FormlyFormOptionsCache } from '../components/formly.field.config';
 import { Subject } from 'rxjs';
+import { defineHiddenProp } from '../utils';
 
 @Injectable({ providedIn: 'root' })
 export class FormlyFormBuilder {
-  constructor(private formlyConfig: FormlyConfig, private componentFactoryResolver: ComponentFactoryResolver) {}
+  constructor(
+    private formlyConfig: FormlyConfig,
+    private componentFactoryResolver: ComponentFactoryResolver,
+    private injector: Injector,
+  ) {}
 
   buildForm(formControl: FormGroup | FormArray, fieldGroup: FormlyFieldConfig[] = [], model: any, options: FormlyFormOptions) {
-    options = options || {};
-    options.formState = options.formState || {};
-
-    if (!options.showError) {
-      options.showError = this.formlyConfig.extras.showError;
-    }
-
-    if (!options.fieldChanges) {
-      options.fieldChanges = new Subject<FormlyValueChangeEvent>();
-    }
-
-    if (!(<FormlyFormOptionsCache> options)._componentFactoryResolver) {
-      (<FormlyFormOptionsCache> options)._componentFactoryResolver = this.componentFactoryResolver;
-    }
-
-    this._buildForm({ fieldGroup, model, formControl, options });
+    this._buildForm({ fieldGroup, model, formControl, options: this._setOptions(options) });
   }
 
   private _buildForm(field: FormlyFieldConfigCache) {
@@ -40,5 +30,40 @@ export class FormlyFormBuilder {
 
   private getExtensions() {
     return Object.keys(this.formlyConfig.extensions).map(name => this.formlyConfig.extensions[name]);
+  }
+
+  private _setOptions(options: FormlyFormOptionsCache) {
+    options = options || {};
+    options.formState = options.formState || {};
+
+    if (!options.showError) {
+      options.showError = this.formlyConfig.extras.showError;
+    }
+
+    if (!options.fieldChanges) {
+      defineHiddenProp(options, 'fieldChanges', new Subject<FormlyValueChangeEvent>());
+    }
+
+    if (!options._componentFactoryResolver) {
+      defineHiddenProp(options, '_componentFactoryResolver', this.componentFactoryResolver);
+    }
+
+    if (!options._injector) {
+      defineHiddenProp(options, '_injector', this.injector);
+    }
+
+    if (!options._markForCheck) {
+      options._markForCheck = (field) => {
+        if (field._componentRefs) {
+          field._componentRefs.forEach(ref => ref.changeDetectorRef.markForCheck());
+        }
+
+        if (field.fieldGroup) {
+          field.fieldGroup.forEach(f => options._markForCheck(f));
+        }
+      };
+    }
+
+    return options;
   }
 }